COMP: Updated ingestion scripts
